Update video.sh
Added weekly code to video.sh, and tweaked variable names to all caps

+++ src/scripts/video.sh
@@ -1,20 +1,38 @@
 #/bin/bash
 
-# Remove Preview Image
-#sudo rm /home/pi/timelapse/media/pictures/preview.jpg
+# This script uses ffmpeg to create a nightly or weekly video in mp4 format
+# from all jpgs in the pictures folder. If the "weekly" flag is passed to the
+# script, it will first create the video, then archive the existing jpgs into
+# a backup folder.
 
-# Run Python Script
-# sudo python3 /home/pi/timelapse/scripts/video.py
+# A nightly video (timelapse.mp4) is maintained in the videos directory.  If
+# run in weekly mode, a copy of the nightly video will be timestamped and 
+# maintained in the video directory as well.
 
-pictures="/home/pi/timelapse/media/pictures"
-videos="/home/pi/timelapse/media/videos"
+# paths and variables.  We should move this code into a config file used across
+# all related programs/scripts
 
-cd $videos
+INSTALL_DIR="/home/pi"
+PICTURES=$INSTALL_DIR"/timelapse/media/pictures"
+VIDEOS=$INSTALL_DIR"/timelapse/media/videos"
+BACKUPS=$INSTALL_DIR"/timelapse/media/backups"
+DATESTAMP=$(date -I)
+
+
+cd $VIDEOS
+if [ -f ./timelapse.mp4 ]; then 
 mv timelapse.mp4 old-timelapse.mp4
+fi
 
-cd $pictures
-ffmpeg -hide_banner -loglevel error -pattern_type glob -r 3 -i "*.jpg" -s 820x616 -vcodec libx264 "$videos/timelapse.mp4"
-# TODO: Archive images
+cd $PICTURES
+ffmpeg -hide_banner -loglevel error -pattern_type glob -r 3 -i "*.jpg" -s 820x616 -vcodec libx264 "$VIDEOS/timelapse.mp4"
 
-# cd $videos
-# TODO: combine videos - not need
+
+
+if [[ $1 == "weekly" ]]; then
+  cd $VIDEOS
+  cp timelapse.mp4 "timelapse-"$DATESTAMP".mp4"
+  cd $PICTURES
+  tar -cf $BACKUPS"/pictures-"$DATESTAMP".tar"  *.jpg
+  rm -f *.jpg
+fi
